fix: Resolve all clippy warnings
- Remove duplicate #![cfg(windows)] attributes from module files (already in lib.rs)
- Remove unused mut from SECURITY_ATTRIBUTES
- Collapse nested if statements in cancellation.rs and io_pump.rs
- Remove unnecessary casts (i32 -> i32)
- Use matches! macro instead of match in job_manager.rs
- Use .is_null() instead of != null_mut()
- Remove redundant closure in process_launcher.rs
- Remove unnecessary mut from startup_info parameter
- Fix all clippy warnings for cleaner code

+++ src/pipe_server.rs
@@ -1,5 +1,3 @@
-#![cfg(windows)]
-
 use crate::cancellation::CancellationManager;
 use crate::error::{Error, Result};
 use crate::io_pump::{start_output_pump, StreamPump};
@@ -374,7 +372,7 @@ impl PipeServer {
                     Some(&mut bytes_read),
                     None,
                 )
-                .map_err(|e| Error::Io(std::io::Error::from_raw_os_error(e.code().0 as i32)))?;
+                .map_err(|e| Error::Io(std::io::Error::from_raw_os_error(e.code().0)))?;
             }
 
             if bytes_read == 0 {
@@ -416,7 +414,7 @@ impl PipeServer {
                     Some(&mut bytes_written),
                     None,
                 )
-                .map_err(|e| Error::Io(std::io::Error::from_raw_os_error(e.code().0 as i32)))?;
+                .map_err(|e| Error::Io(std::io::Error::from_raw_os_error(e.code().0)))?;
             }
             offset += bytes_written as usize;
         }
